requirements.txt
pandas
SQLAlchemy
python-dotenv
google-generativeai
streamlit
tabulate 